Moving files into `components/uielements/input`
