+++ src/renderer/components/uielements/input/InputBigNumber.stories.tsx
@@ -6,8 +6,8 @@ import { Form, Row } from 'antd'
 import { Rule } from 'antd/lib/form'
 import BigNumber from 'bignumber.js'
 
-import Button from '../../button'
-import InputBigNumber from './InputBigNumber'
+import Button from '../button'
+import { InputBigNumber } from './InputBigNumber'
 
 type FormValues = {
   amount: BigNumber
