+++ src/renderer/components/uielements/input/index.tsx
@@ -1,1 +1,2 @@
 export { Input, InputPassword, InputTextArea, InputNumber } from './Input.style'
+export { InputBigNumber } from './InputBigNumber'
