+++ src/renderer/components/wallet/txs/FreezeForm.tsx
@@ -15,7 +15,7 @@ import { useIntl } from 'react-intl'
 import { BNB_SYMBOL } from '../../../helpers/assetHelper'
 import { sequenceTOption } from '../../../helpers/fpHelpers'
 import { AssetWithBalance, FreezeAction, FreezeTxParams } from '../../../services/binance/types'
-import InputBigNumber from '../../uielements/assets/input/InputBigNumber'
+import { InputBigNumber } from '../../uielements/input'
 import AccountSelector from '../AccountSelector'
 import * as Styled from './Form.style'
 import { validateFreezeInput } from './util'
