+++ src/renderer/components/wallet/txs/SendForm.tsx
@@ -20,8 +20,7 @@ import { getBnbAmount } from '../../../helpers/binanceHelper'
 import * as walletRoutes from '../../../routes/wallet'
 import { SendTxParams } from '../../../services/binance/transaction'
 import { AssetWithBalance, AddressValidation, AssetsWithBalance } from '../../../services/binance/types'
-import InputBigNumber from '../../uielements/assets/input/InputBigNumber'
-import { Input } from '../../uielements/input'
+import { Input, InputBigNumber } from '../../uielements/input'
 import AccountSelector from './../AccountSelector'
 import * as Styled from './Form.style'
 import { sendAmountValidator } from './util'
